Fix data type error in SQL migration: cast SUM results to bigint to maintain original column types

+++ src/app/api/migrate/route.ts
@@ -19,8 +19,8 @@ export async function POST(request: NextRequest) {
     switch (migration) {
       case 'fix_zero_impressions':
         sql = `
-          -- Migration 004: Fix all rollup views to filter out zero impressions
-          -- This updates all rollup views directly to ensure zero impressions are filtered
+          -- Migration 005: Fix zero impressions with correct data types
+          -- This maintains the original bigint data types while filtering zero impressions
 
           -- First, update the campaign_content_clean view
           CREATE OR REPLACE VIEW campaign_content_clean AS
@@ -35,13 +35,13 @@ export async function POST(request: NextRequest) {
           WHERE impression > 0  -- Filter out rows with zero impressions
           GROUP BY campaign_id, content_title, content_network_name;
 
-          -- Update the app rollup view directly
+          -- Update the app rollup view directly with correct data types
           CREATE OR REPLACE VIEW rr_rollup_app AS
           SELECT 
             campaign_id,
             content_network_name as app_name,
-            SUM(total_impressions) as impressions,
-            SUM(total_completes) as completes,
+            SUM(total_impressions)::bigint as impressions,
+            SUM(total_completes)::bigint as completes,
             CASE 
               WHEN SUM(total_impressions) > 0 THEN ROUND((SUM(total_completes) * 100.0 / SUM(total_impressions)), 2)
               ELSE 0 
@@ -52,13 +52,13 @@ export async function POST(request: NextRequest) {
           GROUP BY campaign_id, content_network_name
           ORDER BY campaign_id, impressions DESC;
 
-          -- Update the genre rollup view directly
+          -- Update the genre rollup view directly with correct data types
           CREATE OR REPLACE VIEW rr_rollup_genre AS
           SELECT 
             campaign_id,
             'Unknown' as genre_canon,
-            SUM(total_impressions) as impressions,
-            SUM(total_completes) as completes,
+            SUM(total_impressions)::bigint as impressions,
+            SUM(total_completes)::bigint as completes,
             CASE 
               WHEN SUM(total_impressions) > 0 THEN ROUND((SUM(total_completes) * 100.0 / SUM(total_impressions)), 2)
               ELSE 0 
@@ -69,15 +69,15 @@ export async function POST(request: NextRequest) {
           GROUP BY campaign_id
           ORDER BY campaign_id, impressions DESC;
 
-          -- Update the content rollup view directly
+          -- Update the content rollup view directly with correct data types
           CREATE OR REPLACE VIEW rr_rollup_content AS
           SELECT 
             campaign_id,
             content_title as content_key,
             content_title,
             content_network_name,
-            SUM(total_impressions) as impressions,
-            SUM(total_completes) as completes,
+            SUM(total_impressions)::bigint as impressions,
+            SUM(total_completes)::bigint as completes,
             CASE 
               WHEN SUM(total_impressions) > 0 THEN ROUND((SUM(total_completes) * 100.0 / SUM(total_impressions)), 2)
               ELSE 0
